+++ src/cmd/cache/cache.go
@@ -2,6 +2,7 @@ package cmdlipcache
 
 import (
 	"flag"
+	"os"
 
 	cmdlipcachepurge "github.com/liteldev/lip/cmd/cache/purge"
 	"github.com/liteldev/lip/utils/logger"
@@ -25,15 +26,6 @@ Options:
 
 // Run is the entry point.
 func Run(args []string) {
-	// If there is a subcommand, run it and exit.
-	if len(args) >= 1 {
-		switch args[0] {
-		case "purge":
-			cmdlipcachepurge.Run(args[1:])
-			return
-		}
-	}
-
 	flagSet := flag.NewFlagSet("cache", flag.ExitOnError)
 
 	// Rewrite the default usage message.
@@ -42,16 +34,26 @@ func Run(args []string) {
 	}
 
 	var flagDict FlagDict
-
 	flagSet.BoolVar(&flagDict.helpFlag, "help", false, "")
 	flagSet.BoolVar(&flagDict.helpFlag, "h", false, "")
-
 	flagSet.Parse(args)
 
 	// Help flag has the highest priority.
 	if flagDict.helpFlag {
 		logger.Info(helpMessage)
 		return
+	}
+
+	// If there is a subcommand, run it and exit.
+	if flagSet.NArg() >= 1 {
+		switch flagSet.Arg(0) {
+		case "purge":
+			cmdlipcachepurge.Run(args[1:])
+			return
+		default:
+			logger.Error("Unknown command.")
+			os.Exit(1)
+		}
 	}
 
 	// If there is no subcommand, print help message and exit.