+++ src/cmd/install/install.go
@@ -7,6 +7,7 @@ import (
 	"path/filepath"
 
 	cmdlipuninstall "github.com/liteldev/lip/cmd/uninstall"
+	"github.com/liteldev/lip/download"
 	"github.com/liteldev/lip/localfile"
 	"github.com/liteldev/lip/specifiers"
 	"github.com/liteldev/lip/tooth/toothfile"
@@ -18,16 +19,17 @@ import (
 
 // FlagDict is a dictionary of flags.
 type FlagDict struct {
-	helpFlag           bool
-	upgradeFlag        bool
-	forceReinstallFlag bool
-	yesFlag            bool
+	helpFlag            bool
+	upgradeFlag         bool
+	forceReinstallFlag  bool
+	yesFlag             bool
+	numericProgressFlag bool
+	noDependenciesFlag  bool
 }
 
 const helpMessage = `
 Usage:
-  lip install [options] <requirement specifiers>
-  lip install [options] <tooth url/files>
+  lip install [options] <specifiers>
 
 Description:
   Install tooths from:
@@ -39,17 +41,13 @@ Options:
   -h, --help                  Show help.
   --upgrade                   Upgrade the specified tooth to the newest available version.
   --force-reinstall           Reinstall the tooth even if they are already up-to-date.
-  -y, --yes                   Assume yes to all prompts and run non-interactively.`
+  -y, --yes                   Assume yes to all prompts and run non-interactively.
+  --numeric-progress          Show numeric progress instead of progress bar.
+  --no-dependencies            Do not install dependencies.`
 
 // Run is the entry point.
 func Run(args []string) {
 	var err error
-
-	// If there is no argument, print help message and exit.
-	if len(args) == 0 {
-		logger.Info(helpMessage)
-		return
-	}
 
 	flagSet := flag.NewFlagSet("install", flag.ExitOnError)
 
@@ -59,17 +57,14 @@ func Run(args []string) {
 	}
 
 	var flagDict FlagDict
-
 	flagSet.BoolVar(&flagDict.helpFlag, "help", false, "")
 	flagSet.BoolVar(&flagDict.helpFlag, "h", false, "")
-
 	flagSet.BoolVar(&flagDict.upgradeFlag, "upgrade", false, "")
-
 	flagSet.BoolVar(&flagDict.forceReinstallFlag, "force-reinstall", false, "")
-
 	flagSet.BoolVar(&flagDict.yesFlag, "yes", false, "")
 	flagSet.BoolVar(&flagDict.yesFlag, "y", false, "")
-
+	flagSet.BoolVar(&flagDict.numericProgressFlag, "numeric-progress", false, "")
+	flagSet.BoolVar(&flagDict.noDependenciesFlag, "no-dependencies", false, "")
 	flagSet.Parse(args)
 
 	// Help flag has the highest priority.
@@ -78,29 +73,31 @@ func Run(args []string) {
 		return
 	}
 
+	// At least one argument is required.
+	if flagSet.NArg() == 0 {
+		logger.Error("Too few arguments")
+		os.Exit(1)
+	}
+
 	// 1. Validate the requirement specifier or tooth url/path.
 	//    This process will check if the tooth file exists or the tooth url can be accessed
 	//    and if the requirement specifier syntax is valid. For requirement specifier, it
 	//    will also check if the tooth repository can be accessed via GOPROXY.
 
-	logger.Info("Validating requirement specifiers and tooth url/files...")
+	logger.Info("Validating specifiers...")
 
 	// Make specifierList.
 	var specifierList []specifiers.Specifier
 	for _, specifierString := range flagSet.Args() {
+		logger.Info("  Validating " + specifierString + "...")
+
 		specifier, err := specifiers.New(specifierString)
 		if err != nil {
 			logger.Error(err.Error())
-			return
+			os.Exit(1)
 		}
 
 		specifierList = append(specifierList, specifier)
-	}
-
-	// Check if the requirement specifier or tooth url/path is missing.
-	if len(specifierList) == 0 {
-		logger.Error("missing requirement specifier or tooth url/path")
-		return
 	}
 
 	// 2. Parse dependency and download tooth files.
@@ -110,7 +107,7 @@ func Run(args []string) {
 	//    If it is not downloaded, it will be parsed to get its dependencies and add them
 	//    to the array. This process will continue until the array is empty.
 
-	logger.Info("Parsing dependencies and downloading tooth files...")
+	logger.Info("Resolving dependencies and downloading tooths...")
 
 	// An array of downloaded tooth files.
 	// The key is the specifier and the value is the path of the downloaded tooth file.
@@ -135,14 +132,23 @@ func Run(args []string) {
 
 		logger.Info("  Fetching " + specifier.String() + "...")
 
+		var progressBarStyle download.ProgressBarStyleType
+		if logger.GetLevel() > logger.InfoLevel {
+			progressBarStyle = download.StyleNone
+		} else if flagDict.numericProgressFlag {
+			progressBarStyle = download.StylePercentageOnly
+		} else {
+			progressBarStyle = download.StyleDefault
+		}
+
 		// Get tooth file
-		isCached, downloadedToothFilePath, err := getTooth(specifier)
+		isCached, downloadedToothFilePath, err := getTooth(specifier, progressBarStyle)
 		if err != nil {
 			logger.Error(err.Error())
-			return
+			os.Exit(1)
 		}
 		if isCached {
-			logger.Info("    The tooth file is already cached.")
+			logger.Info("    Cached.")
 		}
 
 		// Add the downloaded path to the downloaded tooth files.
@@ -152,57 +158,64 @@ func Run(args []string) {
 		toothFile, err := toothfile.New(downloadedToothFilePath)
 		if err != nil {
 			logger.Error(err.Error())
-			return
+			os.Exit(1)
 		}
 
 		// Validate the tooth file.
 		toothPath := toothFile.Metadata().ToothPath
-		if specifier.Type() == specifiers.RequirementSpecifierType &&
+		if specifier.Type() == specifiers.RequirementKind &&
 			toothPath != specifier.ToothRepo() {
-			logger.Error("the tooth path of the downloaded tooth file does not match the requirement specifier")
+			logger.Error("The tooth path of the downloaded tooth file does not match the requirement specifier")
 
 			// Remove the downloaded tooth file.
 			err = os.Remove(downloadedToothFilePath)
 			if err != nil {
-				logger.Error("failed to remove the downloaded tooth file: " + err.Error())
+				logger.Error("Failed to remove the downloaded tooth file: " + err.Error())
 			}
-			return
+			os.Exit(1)
 		}
 
-		dependencies := toothFile.Metadata().Dependencies
+		// If the no-dependencies flag is set, skip.
+		if !flagDict.noDependenciesFlag {
+			logger.Info("    Resolving dependencies...")
 
-		// Get proper version of each dependency and add them to the queue.
-		for toothPath, versionRange := range dependencies {
-			versionList, err := toothrepo.FetchVersionList(toothPath)
-			if err != nil {
-				logger.Error(err.Error())
-				return
-			}
+			dependencies := toothFile.Metadata().Dependencies
 
-			isMatched := false
-		selectVersion:
-			for _, version := range versionList {
-				for _, innerVersionRange := range versionRange {
-					for _, versionMatch := range innerVersionRange {
-						if versionMatch.Match(version) {
-							// Add the specifier to the queue.
-							specifier, err := specifiers.New(toothPath + "@" + version.String())
-							if err != nil {
-								logger.Error(err.Error())
-								return
+			// Get proper version of each dependency and add them to the queue.
+			for toothPath, versionRange := range dependencies {
+				logger.Info("      Resolving " + toothPath + "...")
+
+				versionList, err := toothrepo.FetchVersionList(toothPath)
+				if err != nil {
+					logger.Error(err.Error())
+					os.Exit(1)
+				}
+
+				isMatched := false
+			selectVersion:
+				for _, version := range versionList {
+					for _, innerVersionRange := range versionRange {
+						for _, versionMatch := range innerVersionRange {
+							if versionMatch.Match(version) {
+								// Add the specifier to the queue.
+								specifier, err := specifiers.New(toothPath + "@" + version.String())
+								if err != nil {
+									logger.Error(err.Error())
+									os.Exit(1)
+								}
+								specifiersToFetch.PushBack(specifier)
+								isMatched = true
+								break selectVersion
 							}
-							specifiersToFetch.PushBack(specifier)
-							isMatched = true
-							break selectVersion
 						}
 					}
 				}
-			}
 
-			if !isMatched {
-				// If no version is selected, error.
-				logger.Error("no version of " + toothPath + " matches the requirement of " + specifier.String())
-				return
+				if !isMatched {
+					// If no version is selected, error.
+					logger.Error("No version of " + toothPath + " matches the requirement of " + specifier.String())
+					os.Exit(1)
+				}
 			}
 		}
 	}
@@ -216,29 +229,33 @@ func Run(args []string) {
 
 	if flagDict.forceReinstallFlag || flagDict.upgradeFlag {
 		if flagDict.forceReinstallFlag {
-			logger.Info("force-reinstall flag is set, reinstalling all installed tooth specified by the specifiers...")
+			logger.Info("force-reinstall flag is set, Lip will reinstall all installed tooths specified by the specifiers...")
 		} else {
-			logger.Info("upgrade flag is set, upgrading all installed tooth specified by the specifiers...")
+			logger.Info("upgrade flag is set, Lip will upgrade all installed tooths specified by the specifiers...")
 		}
 
+		logger.Info("Uninstalling tooths to be reinstalled or upgraded...")
+
 		for _, specifier := range specifierList {
+			logger.Info("  Resolving " + specifier.String() + "...")
+
 			// If the specifier is not a requirement specifier, skip.
-			if specifier.Type() != specifiers.RequirementSpecifierType {
-				logger.Error("the specifier " + specifier.String() + " is not a requirement specifier. It cannot be used with the force-reinstall flag or the upgrade flag")
+			if specifier.Type() != specifiers.RequirementKind {
+				logger.Error("The specifier " + specifier.String() + " is not a requirement specifier. It cannot be used with the force-reinstall flag or the upgrade flag")
 				continue
 			}
 
 			toothFile, err := toothfile.New(downloadedToothFiles[specifier.String()])
 			if err != nil {
 				logger.Error(err.Error())
-				return
+				os.Exit(1)
 			}
 
 			// If the tooth file of the specifier is not installed, skip.
 			isInstalled, err := toothrecord.IsToothInstalled(toothFile.Metadata().ToothPath)
 			if err != nil {
 				logger.Error(err.Error())
-				return
+				os.Exit(1)
 			}
 			if !isInstalled {
 				continue
@@ -248,14 +265,14 @@ func Run(args []string) {
 			recordDir, err := localfile.RecordDir()
 			if err != nil {
 				logger.Error(err.Error())
-				return
+				os.Exit(1)
 			}
 
 			toothRecordFilePath := filepath.Join(recordDir, localfile.GetRecordFileName(toothFile.Metadata().ToothPath))
 			toothRecord, err := toothrecord.New(toothRecordFilePath)
 			if err != nil {
 				logger.Error(err.Error())
-				return
+				os.Exit(1)
 			}
 
 			// Compare the version of the tooth file and the version of the tooth record.
@@ -266,15 +283,15 @@ func Run(args []string) {
 			}
 
 			// If the tooth file of the specifier is installed, uninstall it.
-			logger.Info("  Uninstalling " + toothFile.Metadata().ToothPath + "...")
+			logger.Info("    Uninstalling " + toothFile.Metadata().ToothPath + "...")
 
 			possessionList := toothFile.Metadata().Possession
 			recordFileName := localfile.GetRecordFileName(toothFile.Metadata().ToothPath)
 
-			err = cmdlipuninstall.Uninstall(recordFileName, possessionList)
+			err = cmdlipuninstall.Uninstall(recordFileName, possessionList, flagDict.yesFlag)
 			if err != nil {
 				logger.Error(err.Error())
-				return
+				os.Exit(1)
 			}
 		}
 
@@ -286,15 +303,15 @@ func Run(args []string) {
 	//    will be installed. If the tooth file is installed but the version is different,
 	//    it will be upgraded or reinstalled according to the flags.
 
-	logger.Info("Installing tooth files...")
+	logger.Info("Installing tooths...")
 
 	// Store downloaded tooth files in an array.
 	downloadedToothFileList := make([]toothfile.ToothFile, 0)
 	for _, downloadedToothFilePath := range downloadedToothFiles {
 		toothFile, err := toothfile.New(downloadedToothFilePath)
 		if err != nil {
-			logger.Error("failed to open the downloaded tooth file " + downloadedToothFilePath + ": " + err.Error())
-			return
+			logger.Error("Failed to open the downloaded tooth file " + downloadedToothFilePath + ": " + err.Error())
+			os.Exit(1)
 		}
 
 		downloadedToothFileList = append(downloadedToothFileList, toothFile)
@@ -303,33 +320,35 @@ func Run(args []string) {
 	// Topological sort the array in descending order.
 	downloadedToothFileList, err = sortToothFiles(downloadedToothFileList)
 	if err != nil {
-		logger.Error("failed to sort the downloaded tooth files: " + err.Error())
-		return
+		logger.Error("Failed to sort the downloaded tooth files: " + err.Error())
+		os.Exit(1)
 	}
 
 	for _, toothFile := range downloadedToothFileList {
+		logger.Info("  Resolving " + toothFile.Metadata().ToothPath + "@" + toothFile.Metadata().Version.String() + "...")
+
 		// If the tooth file is already installed, skip.
 		isInstalled, err := toothrecord.IsToothInstalled(toothFile.Metadata().ToothPath)
 		if err != nil {
 			logger.Error(err.Error())
-			return
+			os.Exit(1)
 		}
 		if isInstalled {
-			logger.Info("  " + toothFile.Metadata().ToothPath + " is already installed.")
+			logger.Info("    " + toothFile.Metadata().ToothPath + " is already installed.")
 			continue
 		}
 
 		// Install the tooth file.
-		logger.Info("  Installing " + toothFile.Metadata().ToothPath + "@" +
+		logger.Info("    Installing " + toothFile.Metadata().ToothPath + "@" +
 			toothFile.Metadata().Version.String() + "...")
 
 		// TODO: Check if the tooth file is manually installed.
 		isManuallyInstalled := false
 
-		err = Install(toothFile, isManuallyInstalled, flagDict.yesFlag)
+		err = install(toothFile, isManuallyInstalled, flagDict.yesFlag)
 		if err != nil {
 			logger.Error(err.Error())
-			return
+			os.Exit(1)
 		}
 	}